+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+## [0.2.1]
+- easy api improvements
+- adjusted visualization
+- bugfixes in velocity and acceleration visualization
+- optional consideration of traffic lights and stop lines
+
+
 ## [0.1.1]
 - added cr reactive planner integration via method for global trajectory
 - support of commonroad-route-planner 2025.0.1 api changes